+++ dot/config_test.go
@@ -2,25 +2,3 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 
 package dot
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-// TestExportConfig tests exporting a toml configuration file
-func TestExportConfig(t *testing.T) {
-	cfg, cfgFile := newTestConfigWithFile(t)
-	require.NotNil(t, cfg)
-
-	genFile := NewTestGenesisRawFile(t, cfg)
-
-	cfg.Init.Genesis = genFile
-
-	err := InitNode(cfg)
-	require.Nil(t, err)
-
-	file := exportConfig(cfg, cfgFile.Name())
-	require.NotNil(t, file)
-}
